requirements.txt
langchain>=0.1
langchain-community
langchain-google-genai
chromadb
pandas
pyyaml
python-dotenv
gradio
